LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/the-stylepropertymap/properties/flex-expected.txt

CONSOLE MESSAGE: ReferenceError: Can't find variable: CSSPositionValue

Harness Error (FAIL), message = ReferenceError: Can't find variable: CSSPositionValue
